Update to Go 1.22, Alpine 3.20
This also embeds version numbers directly in the template so they're easier to maintain (since we're not scraping them from anywhere, `versions.json` doesn't add a lot of value here, just overhead).

I chose Go 1.22 instead of 1.23 as a temporary measure because of https://go.dev/issues/68976 (and the potential of that to make this harder to test).